requirements.txt
streamlit==1.24.0
#pydantic>=1.4.0,<2
#fschat==0.2.18
GPUtil==1.4.0
psutil==5.9.5
py-cpuinfo==9.0.0
